A number of refinements, formatting and fixes to bring steps 1 .. 8 to a reviewable state
Relates to issue #7 and #16
