+++ src/step-2/routes/users.js
@@ -1,7 +1,6 @@
 export default async function users(fastify) {
-  fastify.get(
-    '/users',
-    {},
-    async () => [{ username: 'alice' }, { username: 'bob' }]
-  )
+  fastify.get('/users', {}, async () => [
+    { username: 'alice' },
+    { username: 'bob' },
+  ])
 }
