+++ src/step-6/test/index.test.js
@@ -1,13 +1,12 @@
 import t from 'tap'
 
 import buildServer from '../index.js'
-import config from '../config.js'
 
 const { test } = t
 
 test('GET /users', async t => {
   t.test('returns users', async t => {
-    const fastify = buildServer(config)
+    const fastify = buildServer()
 
     const res = await fastify.inject('/users')
