+++ src/step-6/test/login.test.js
@@ -1,60 +1,31 @@
 import t from 'tap'
 import fastify from 'fastify'
-import sinon from 'sinon'
 
 const { test } = t
 
 function buildServer() {
   return fastify()
-    .decorate('jwt', { sign: sinon.stub() })
     .register(import('../routes/login.js'))
 }
 
 test('POST /login', async t => {
-  t.test('returns 400 with missing credentials', async t => {
-    const fastify = buildServer()
-
-    const res = await fastify.inject({
-      url: '/login',
-      method: 'POST',
-    })
-
-    t.strictEqual(res.statusCode, 400)
-  })
-
-  t.test('returns 400 with partial credentials', async t => {
+  t.test('returns 400 with an invalid post payload', async t => {
     const fastify = buildServer()
 
     const res = await fastify.inject({
       url: '/login',
       method: 'POST',
       body: {
-        username: 'alice',
+        name: 'alice',
+        passcode: 'alice',
       },
     })
 
     t.strictEqual(res.statusCode, 400)
   })
 
-  t.test('returns 401 with wrong credentials', async t => {
+  t.test('returns the data with valid post payload', async t => {
     const fastify = buildServer()
-
-    const res = await fastify.inject({
-      url: '/login',
-      method: 'POST',
-      body: {
-        username: 'alice',
-        password: 'wrong password',
-      },
-    })
-
-    t.strictEqual(res.statusCode, 401)
-  })
-
-  t.test('obtains a token with right credentials', async t => {
-    const fastify = buildServer()
-
-    fastify.jwt.sign.returns('jwt token')
 
     const res = await fastify.inject({
       url: '/login',
@@ -66,6 +37,9 @@ test('POST /login', async t => {
     })
 
     t.strictEqual(res.statusCode, 200)
-    t.strictEqual((await res.json()).token, 'jwt token')
+    t.equivalent(await res.json(), {
+      username: 'alice',
+      password: 'alice',
+    })
   })
 })
